# app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Briva API",
    "description": "Briva - Yapay Zekâ Destekli Gönüllülük Platformu API Dokümantasyonu",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000",
      "description": "Lokal Geliştirme Sunucusu"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Yeni kullanıcı kaydı",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "Sifre123_!"
                  },
                  "role": {
                    "type": "string",
                    "example": "volunteer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Kayıt başarılı"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Kullanıcı girişi",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "Sifre123_!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Giriş başarılı ve JWT döner"
          }
        }
      }
    },
    "/api/events": {
      "get": {
        "summary": "Etkinlikleri listele",
        "tags": [
          "Events"
        ],
        "security": [],
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Şehir filtresi"
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Kategori filtresi"
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Anahtar kelime araması: başlık, açıklama ve gereksinimlerde arar (Sprint 2)"
          },
          {
            "name": "start_after",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Bu tarihten sonra başlayan etkinlikler (ISO 8601)"
          },
          {
            "name": "start_before",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Bu tarihten önce başlayan etkinlikler (ISO 8601)"
          }
        ],
        "responses": {
          "200": {
            "description": "Etkinlik listesi"
          }
        }
      },
      "post": {
        "summary": "Yeni etkinlik oluştur (STK)",
        "tags": [
          "Events"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Doğa Temizliği"
                  },
                  "start_date": {
                    "type": "string",
                    "example": "2026-08-01T10:00:00"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Etkinlik oluşturuldu"
          }
        },
        "description": "Issue #18: yanıt, açıklama kalitesini ölçen ai_analysis (score, warnings, suggestions) alanını da içerir."
      }
    },
    "/api/recommendations": {
      "post": {
        "summary": "Etkinlik önerisi al",
        "tags": [
          "Recommendations"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "city": {
                    "type": "string",
                    "example": "İstanbul"
                  },
                  "interests": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "çevre"
                    ]
                  },
                  "skills": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "organizasyon"
                    ]
                  },
                  "available_days": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "saturday"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Öneri listesi"
          }
        }
      }
    },
    "/api/favorites": {
      "get": {
        "summary": "Favori etkinliklerimi listele (Sprint 2)",
        "tags": [
          "Favoriler"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Favori listesi"
          }
        }
      }
    },
    "/api/favorites/{event_id}": {
      "post": {
        "summary": "Etkinliği favorilere ekle (Sprint 2)",
        "tags": [
          "Favoriler"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Favorilere eklendi"
          },
          "409": {
            "description": "Zaten favorilerde"
          }
        }
      },
      "delete": {
        "summary": "Etkinliği favorilerden çıkar (Sprint 2)",
        "tags": [
          "Favoriler"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Favorilerden çıkarıldı"
          }
        }
      }
    },
    "/api/ai/analyze-event": {
      "post": {
        "summary": "AI Etkinlik Analizörü — ilan kalitesini analiz et (Sprint 2)",
        "description": "event_id ile mevcut bir etkinliği ya da title/description ile taslak bir ilanı analiz eder. Gemini API kullanır; anahtar yoksa kural tabanlı fallback devrededir.",
        "tags": [
          "AI"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "event_id": {
                    "type": "integer"
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "requirements": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Analiz sonucu: quality_score, strengths, weaknesses, suggestions, improved_description"
          }
        }
      }
    },
    "/api/feedback/{event_id}": {
      "post": {
        "summary": "Öneriye geri bildirim ver: like/dislike (Sprint 2)",
        "tags": [
          "Geri Bildirim"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "feedback": {
                    "type": "string",
                    "enum": [
                      "like",
                      "dislike"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Kaydedildi"
          },
          "200": {
            "description": "Güncellendi"
          }
        }
      },
      "delete": {
        "summary": "Geri bildirimi kaldır (Sprint 2)",
        "tags": [
          "Geri Bildirim"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Kaldırıldı"
          }
        }
      }
    },
    "/api/feedback/my": {
      "get": {
        "summary": "Geri bildirimlerimi listele (Sprint 2)",
        "tags": [
          "Geri Bildirim"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Liste"
          }
        }
      }
    },
    "/api/notifications": {
      "get": {
        "summary": "Bildirimlerimi listele (Sprint 2)",
        "tags": [
          "Bildirimler"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "unread",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "true ise sadece okunmamışlar"
          }
        ],
        "responses": {
          "200": {
            "description": "Bildirim listesi + unread_count"
          }
        }
      }
    },
    "/api/notifications/{notification_id}/read": {
      "put": {
        "summary": "Bildirimi okundu işaretle (Sprint 2)",
        "tags": [
          "Bildirimler"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "İşaretlendi"
          }
        }
      }
    },
    "/api/notifications/read-all": {
      "put": {
        "summary": "Tüm bildirimleri okundu işaretle (Sprint 2)",
        "tags": [
          "Bildirimler"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "İşaretlendi"
          }
        }
      }
    }
  }
}